requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.7
requests==2.25.1
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
urllib3==1.26.4
youtube-dl==2021.4.26
zope.interface==5.4.0
